refactor: move LinkUserInfo type to Onramp.ts and update imports
